python-trading-application-main/requirements.txt
yfinance
mysql-connector-python
pandas
requests
ta
matplotlib
SQLAlchemy
numpy